scratchwork for reorganization (don't merge)
